+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/StackSizes.java
@@ -35,7 +35,7 @@ public class StackSizes extends Feature {
     public static Boolean foodStackReduction = true;
     @Config
     @Label(name = "Food Stack Reduction Formula", description = "The formula to calculate the stack size of a food item. Variables as hunger, saturation_modifier, effectiveness as numbers and fast_food as boolean can be used. This is evaluated with EvalEx https://ezylang.github.io/EvalEx/concepts/parsing_evaluation.html.")
-    public static String foodStackReductionFormula = "ROUND((1 - (effectiveness - 1) / 24) * 64 * 0.25, 0)";
+    public static String foodStackReductionFormula = "ROUND((1 - (effectiveness - 1) / 25) * 64 * 0.25, 0)";
     @Config(min = 1, max = 64)
     @Label(name = "Stackable Stews", description = "Stews will stack up to this number. It's overridden by 'foodStackReduction' if enabled. Still affected by black/whitelist")
     public static Integer stackableSoups = 16;